Add files via upload
Working with the COVID19 dataset, published by John Hopkins University, which consists of the data related to the cumulative number of confirmed cases per day in each Country. Also, we have another dataset consist of various life factors, scored by the people living in each country around the globe. Merging these two datasets to see if there is any relationship between the spread of the virus in a country and how happy people are, living in that country.